+++ src/pages/Login.jsx
@@ -1,7 +1,27 @@
-// import React from "react";
+import React from "react";
+import Navbar from "../components/Navbar";
+import "../App.css";
 
-function Login() {
-  return <h1>Login</h1>;
-}
+const Login = () => {
+    return(
+        <div className="wrapper">
+          <Navbar/>
+            <form action="">
+                <h1>Hello !</h1>
+                <h2>Come Make We Work</h2>
+                <div className="input-box">
+                    <input type="text" placeholder="Username" required/>
+                </div>
+                <div className="input-box">
+                    <input type="password" placeholder="password" required />
+                </div>
+                <div className="forgot-password">
+                    <a href="#">Forgot password?</a>
+                </div>
+                <button type="submit">Login</button>
+            </form>
+        </div>
+    );
+};
 
 export default Login;
